requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
db-dtypes==1.2.0
Flask==3.0.3
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pandas-gbq==0.23.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
python-dotenv==1.0.1